fix: fsd 구조에 맞게 컴포넌트 위치 이동
