+++ apps/web/src/features/all-tags/index.ts
@@ -1,0 +1,1 @@
+export { AllTags } from './ui';
