+++ apps/web/src/features/tag-list-item/index.ts
@@ -1,0 +1,1 @@
+export { TagListItem } from './ui';
